Revert some changes

+++ src/include/btree_cmp_inline.h
@@ -307,7 +307,7 @@ __lex_compare_skip_gt_16(
   const uint8_t *ustartp, const uint8_t *tstartp, size_t len, int lencmp, size_t *matchp)
 {
     __m128i res_eq, res_gt, t, u;
-    uint32_t eq_bits;
+    uint16_t eq_bits;
     size_t match;
     const uint8_t *tendp, *treep, *uendp, *userp;
 
@@ -324,7 +324,7 @@ __lex_compare_skip_gt_16(
         u = _mm_loadu_si128((const __m128i *)userp);
         t = _mm_loadu_si128((const __m128i *)treep);
         res_eq = _mm_cmpeq_epi8(u, t);
-        if ((eq_bits = _mm_movemask_epi8(res_eq)) != 65535) {
+        if ((eq_bits = (uint16_t)_mm_movemask_epi8(res_eq)) != 65535) {
             match = (size_t)(userp - ustartp);
             goto final128;
         }
@@ -337,7 +337,7 @@ __lex_compare_skip_gt_16(
     u = _mm_loadu_si128((const __m128i *)(uendp - WT_VECTOR_SIZE));
     t = _mm_loadu_si128((const __m128i *)(tendp - WT_VECTOR_SIZE));
     res_eq = _mm_cmpeq_epi8(u, t);
-    eq_bits = _mm_movemask_epi8(res_eq);
+    eq_bits = (uint16_t)_mm_movemask_epi8(res_eq);
 
 final128:
     if (eq_bits == 65535) {
